backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-filter==23.2
gunicorn==21.2.0
idna==3.4
kombu==5.3.1
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
responses==0.23.3
six==1.16.0
sqlparse==0.4.4
types-PyYAML==6.0.12.11
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
